Add access to class fields
Precalculate the offset of the fields relative to the instance base
address.
Start with the field of the superclass and add own fields afterwards.

Assigning values to class fields now properly set the value in the
object heap.
Class fields are read from the heap as well, when used as a rval.