Add method call support on objects by generating virtual method tables
Generate virtual table with all methods for each class.
Make sure to replace overridden ones in the table.

A list of all vtables is passed as an implicit parameter next to the
|this| pointer. The correct vtable is grabbed and stored in the |this|
structure on object allocation.
Piglet doesn't allow for global variables or other means to access heap
addresses from other procedures :(