Implement SPiglet to Kanga translation
All temps are mapped to stack indices by default, except for the first 4
arguments of a procedure, which are saved in the special a0 - a3
argument registers.

Whenever a temp is used, its value is first loaded from the stack and
written back to the stack after changes.
The temporary registers t0 - t2 are used to cache values before writing
them back to the stack.

Before calling a function the current argument registers a0 - a3 are
saved on the stack and restored afterwards.