Add use, def and flow support for RetExp
Check on every def(pos), use(pos) or flow(pos) in a procedure, if pos ==
position of RetExp, and give back the corresponding use/def/flow for the
RetExp.
If a statements is in a procedure, the statements nextPos() can give the
position of the RetExp, so when running the graph, we arrive at the
RetExp position in the end.
The printer now also prints the RetExp and its use/def/flow.