Reuse dead temp index in following statements
When iterating through all statements in a procedure, remember which
temps were dead already. See if the def() set of the current statement
contains a temp with an higher index than one of the dead ones and
replace the higher temp number with the dead one in all the following
statements.

If a dead temp is written to again before it is reused, remove it from
the list of available dead temp indices.

Repeat the liveness analysis on the changed AST afterwards until no dead
temps are reused anymore.